Split up components into files
